fix dto

+++ stats/stats-dto/src/main/java/ru/practicum/dto/ViewStatsDto.java
@@ -12,6 +12,6 @@ public class ViewStatsDto {
 
     private String app;
     private String uri;
-    private int hits;
+    private long hits;
 
 }
